pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "langgraph-app"
version = "0.1.0"
description = "A LangGraph AI agent application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=1.0.0",
    "langchain-core>=0.3.64",
    "langchain-openai>=1.1.9",
    "python-dotenv>=0.8.0",
    "httpx>=0.28.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pymongo>=4.9.0",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "sqlalchemy[asyncio]>=2.0.47",
    "structlog>=25.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
